chore: Extract requestIdleCallback shim to it's own file (#51212)
https://developer.mozilla.org/en-US/docs/Web/API/Window/requestIdleCallback
